+++ examples/009-clean_architecture_todo_mvc_cloud_firestore_immutable_state/lib/app.dart
@@ -37,7 +37,6 @@ class App extends StatelessWidget {
           ),
         )
       ],
-      reinjectOn: [RM.get<AuthState>()],
       builder: (_) => MaterialApp(
         title: StatesRebuilderLocalizations().appTitle,
         theme: ArchSampleTheme.theme,